+++ RELEASE_NOTES.md
@@ -1,6 +1,8 @@
 ### Version 1.5.0
 __Changes__
 - Added Build Microbial SpeciesTree App
+- Added Localize Domain Annotations App
+
 
 ### Version 1.4.0
 __Changes__